requirements.txt
Django
djangorestframework
dependency_injector
